新增间接头文件引用case

+++ tests/resources_cpp/exports/indirectRef/main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-    int res = funcDoSomething(1, 2);
+    double res = funcDoSomething(1, 2);
     cout << "result is : " << res << endl;
     return 0;
 }
